// frontend/src/api/model/githubCom4H1RZooraInternalDomainMyExamState.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainMyExamState = typeof GithubCom4H1RZooraInternalDomainMyExamState[keyof typeof GithubCom4H1RZooraInternalDomainMyExamState];


export const GithubCom4H1RZooraInternalDomainMyExamState = {
  MyExamStateUpcoming: 'upcoming',
  MyExamStateOpen: 'open',
  MyExamStateSubmitted: 'submitted',
  MyExamStateGraded: 'graded',
} as const;